+++ lib/src/powersync_impl.dart
@@ -3,6 +3,7 @@ import 'package:sqlite_async/sqlite_async.dart';
 import 'package:sqlite3/sqlite3.dart' as sqlite;
 import 'package:meta/meta.dart';
 import 'core.dart';
+import 'utils.dart';
 
 export 'core.dart';