+++ lib/src/postgres_impl.dart
@@ -74,6 +74,7 @@ class PostgresWriteContext extends PostgresReadContext implements SqlRecords {
   @override
   Future<MutationResult> execute<P>(Command<P> mutation, [P? params]) async {
     final (sql, map) = mutation.apply(params);
+    if (sql == NoOpCommand) return const NoOpMutationResult();
     final result = await _session.execute(pg.Sql.named(sql), parameters: map);
     return PostgresMutationResult(result);
   }
@@ -82,6 +83,7 @@ class PostgresWriteContext extends PostgresReadContext implements SqlRecords {
   Future<void> executeBatch<P>(Command<P> mutation, List<P> paramsList) async {
     for (final p in paramsList) {
       final (sql, map) = mutation.apply(p);
+      if (sql == NoOpCommand) continue;
       await _session.execute(pg.Sql.named(sql), parameters: map);
     }
   }